Resolved comments from PR
