+++ src/pages/transactions/components/CreateBankAccountModal.tsx
@@ -13,18 +13,11 @@ import { Dispatch, SetStateAction } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Card, Element } from '@invoiceninja/cards';
 import { Button, InputField } from '@invoiceninja/forms';
-import { AxiosError } from 'axios';
-import { endpoint } from 'common/helpers';
-import { request } from 'common/helpers/request';
-import { toast } from 'common/helpers/toast/toast';
 import { useTitle } from 'common/hooks/useTitle';
 import { BankAccountInput } from 'common/interfaces/bank-accounts';
-import { useState, FormEvent } from 'react';
-import { useQueryClient } from 'react-query';
-
-interface BankAccountValidation {
-  bank_account_name?: string[];
-}
+import { useState } from 'react';
+import { ValidationBag } from 'common/interfaces/validation-bag';
+import { useHandleCreate } from 'pages/settings/bank-accounts/create/hooks/useHandleCreate';
 
 interface Props {
   isModalOpen: boolean;
@@ -36,13 +29,19 @@ export default function CreateBankAccountModal(props: Props) {
 
   useTitle('create_bank_account');
 
-  const queryClient = useQueryClient();
-
   const [isFormBusy, setIsFormBusy] = useState<boolean>(false);
 
-  const [errors, setErrors] = useState<BankAccountValidation>();
+  const [errors, setErrors] = useState<ValidationBag>();
 
   const [bankAccount, setBankAccount] = useState<BankAccountInput>();
+
+  const handleSave = useHandleCreate(
+    bankAccount,
+    setErrors,
+    setIsFormBusy,
+    isFormBusy,
+    props.setIsModalOpen
+  );
 
   const handleChange = (
     property: keyof BankAccountInput,
@@ -57,46 +56,6 @@ export default function CreateBankAccountModal(props: Props) {
     }
   };
 
-  const handleSave = (event: FormEvent<HTMLFormElement>) => {
-    if (!isFormBusy) {
-      event.preventDefault();
-
-      toast.processing();
-      setErrors(undefined);
-      setIsFormBusy(true);
-
-      request('POST', endpoint('/api/v1/bank_integrations'), bankAccount)
-        .then(() => {
-          toast.success('created_bank_account');
-
-          queryClient.invalidateQueries('/api/v1/bank_integrations');
-
-          window.dispatchEvent(
-            new CustomEvent('invalidate.combobox.queries', {
-              detail: {
-                url: endpoint('/api/v1/bank_integrations'),
-              },
-            })
-          );
-
-          props.setIsModalOpen(false);
-        })
-        .catch((error: AxiosError) => {
-          console.error(error);
-
-          if (error?.response?.status === 422) {
-            setErrors(error?.response.data.errors);
-          } else {
-            toast.error();
-          }
-        })
-        .finally(() => {
-          toast.dismiss();
-          setIsFormBusy(false);
-        });
-    }
-  };
-
   return (
     <Modal
       title={t('new_bank_account')}
@@ -106,20 +65,16 @@ export default function CreateBankAccountModal(props: Props) {
       backgroundColor="gray"
     >
       <Card onFormSubmit={handleSave}>
-        <Element leftSide={t('bank_account_name')}>
+        <Element leftSide={t('name')}>
           <InputField
             value={bankAccount?.bank_account_name}
             onValueChange={(value) => handleChange('bank_account_name', value)}
-            errorMessage={errors?.bank_account_name}
+            errorMessage={errors?.errors.bank_account_name}
           />
         </Element>
       </Card>
 
       <div className="flex justify-end space-x-4">
-        <Button type="minimal" onClick={handleCancel}>
-          {t('cancel')}
-        </Button>
-
         <Button onClick={handleSave}>{t('save')}</Button>
       </div>
     </Modal>
