+++ src/pages/transactions/edit/Edit.tsx
@@ -140,8 +140,8 @@ export function Edit() {
         navigate('/transactions');
       })
       .catch((error: AxiosError) => {
-        if (error?.response?.status === 422) {
-          setErrors(error?.response.data.errors);
+        if (error.response?.status === 422) {
+          setErrors(error.response.data.errors);
           toast.dismiss();
         } else {
           console.error(error);
